+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file. This project uses [Semantic Versioning](https://semver.org/)
 
+## v2.2.4 (Pending Release)
+
+### Changed
+
+- fix(ui): bind Location tab actions once
+
 ## v2.2.3 (2026-08-13)
 
 ### Changed